refactor(util): move fold func to util
