+++ index.js
@@ -1,25 +1,5 @@
 const { transfer } = require("./lib/transfer");
-const { escapeData } = require("./lib/util");
-const { walkTree } = require("markmap-common");
-
-
-function fold(tree, maxDepth = 0) {
-  if (
-    typeof maxDepth === 'undefined' ||
-    typeof parseInt(maxDepth) !== 'number' ||
-    maxDepth <= 0
-  ) {
-    return
-  }
-  maxDepth = parseInt(maxDepth)
-  walkTree(tree, (node, next) => {
-    // Here depths are in sequence of [0, 2, 4, ..., 2n]
-    if (node.d >= maxDepth * 2) {
-      node.p = Object.assign(node.p, { f: true });
-    }
-    next();
-  })
-}
+const { escapeData, fold } = require("./lib/util");
 
 hexo.extend.tag.register("markmap", function (args, content) {
   const { svgData } = transfer(content.trim());
