+++ lib/util.js
@@ -1,3 +1,20 @@
+const { walkTree } = require("markmap-common");
+
+function fold(tree, maxDepth = 0) {
+    if (typeof maxDepth === 'undefined' ||
+        typeof parseInt(maxDepth) !== 'number' ||
+        maxDepth <= 0
+    ) { return; }
+    maxDepth = parseInt(maxDepth);
+    walkTree(tree, (node, next) => {
+        // Here depths are in sequence of [0, 2, 4, ..., 2n]
+        if (node.d >= maxDepth * 2) {
+            node.p = Object.assign(node.p, { f: true });
+        }
+        next();
+    });
+}
+
 function escapeData(str) {
     if (str.length == 0) return "";
     let s = str.replace(/&/g, "&amp;");
@@ -9,4 +26,4 @@ function escapeData(str) {
     return s;
 }
 
-module.exports = { escapeData }
+module.exports = { escapeData, fold }
